+++ async/src/test/kotlin/lt/saltyjuice/dragas/chatty/v3/async/mock/controller/NotAsyncController.kt
@@ -4,13 +4,13 @@ import lt.saltyjuice.dragas.chatty.v3.core.route.Controller
 import lt.saltyjuice.dragas.chatty.v3.core.route.On
 import lt.saltyjuice.dragas.chatty.v3.core.route.When
 
-class NotAsyncController : Controller<Int>()
+class NotAsyncController : Controller<Float>()
 {
     @On(Int::class)
     @When("isEven")
     fun onEven(request: Int)
     {
-        writeResponse(request)
+        writeResponse(request.toFloat())
     }
 
     fun isEven(request: Int): Boolean
