+++ async/src/test/kotlin/lt/saltyjuice/dragas/chatty/v3/async/mock/io/MockAdapter.kt
@@ -3,9 +3,9 @@ package lt.saltyjuice.dragas.chatty.v3.async.mock.io
 import lt.saltyjuice.dragas.chatty.v3.core.adapter.Deserializer
 import lt.saltyjuice.dragas.chatty.v3.core.adapter.Serializer
 
-class MockAdapter : Serializer<Int, Int>, Deserializer<Int, Int>
+class MockAdapter : Serializer<Float, Float>, Deserializer<Int, Int>
 {
-    override fun serialize(any: Int): Int
+    override fun serialize(any: Float): Float
     {
         return any
     }
